Move Context into task module
Change-type: patch
