+++ src/lib.rs
@@ -1,8 +1,6 @@
 mod path;
-mod task;
 
 pub mod error;
 pub mod extract;
 pub mod system;
-
-pub use task::*;
+pub mod task;
