+++ tests/library_test.rs
@@ -1,6 +1,6 @@
 use gustav::extract::{Target, Update};
-use gustav::system::{Context, System};
-use gustav::*;
+use gustav::system::System;
+use gustav::task::*;
 use json_patch::Patch;
 use serde_json::{from_value, json};
